Added a preliminary, and rather messy, evolution algorithm based on a mutation rate that changes, adds, or deletes variables from production rules. Special rules are used for the branching variables as they are context sensitive. Their production rules are always identity, and they cannot be replaced or deleted right now. Additionally, the random production rule generator makes sure all branch variables are matched. The fitness function used is the responsiveness of a tree system. The population selection is a weighted random selection, with weights exponentiated to increase the chances of better tree systems. However, it's a bit black-boxy at the moment, the exponentiation factor can be configured but I have no idea how strong or weak it's selection is. I also have no idea how well-chosen the mutation rate is.
I changed the scene camera a tiny bit to get a good look at the 3x3 population of trees. I also tried 4x4 but it can crash Unity (presumably due to a lack of available (video?) memory if the trees get huge).

The result is both promising and kinda lame. I don't know how good my evolutionary algorithm is, as this is the first one I've made and technical details are hard to come by for some reason, but the trees basically refuse to branch, becoming one giant twisty stalk full of leaves. However, I've had several occasions where the leaves would be pretty much maximized along this stalk, with three leaves at every node (this is branching too!), so at least we know the fitness function is working. It's possible additional fitness criteria such as minimized biomass or something could help, but it's also possible the branch handling implementation is just not cutting it. It's also possible the renderer is not well fit enough for the task... There's lots to consider. One particular thing I've noticed is that s small mutation can cause a giant change due to how L-systems work, so it's also possible that the whole thing is just ill-conceived :P